+++ app/collaboration/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import CollaborationLayer from "@/components/collaboration/CollaborationLayer";
+
+export default function CollaborationPage() {
+  return <CollaborationLayer />;
+}